Update batch_scenes2strips.py to version "4" New when building xtrack strips: - Both scene orthos are required. - Scene masks are created considering each ortho separately, then are combined using logical OR to produce the scene mask that is written to source directory. - Two output strip orthos are created; the new ortho is named "*_ortho2.tif"
Fixed a major logical issue in batch_check_setsm.py involving the capability to delete source "check group" files upon encountering an issue with source files.
Now, special check group can be set to "pairnames" with the option to remove source files where missing a check group suffix, and if any of the scenes within that pairname are missing a suffix, files for the whole pairname will be removed.